PatternDefinitionCompiler
