+++ crates/core/src/pattern/pattern_definition.rs
@@ -1,22 +1,17 @@
 use super::{
     patterns::{Matcher, Pattern},
     resolved_pattern::ResolvedPattern,
-    variable::{get_variables, Variable, VariableSourceLocations},
+    variable::Variable,
     State,
 };
-use crate::{
-    context::Context,
-    pattern_compiler::{and_compiler::AndCompiler, CompilationContext, NodeCompiler},
-};
-use anyhow::{anyhow, Result};
+use crate::context::Context;
+use anyhow::Result;
 use marzano_util::analysis_logs::AnalysisLogs;
-use std::collections::BTreeMap;
-use tree_sitter::Node;
 
 #[derive(Clone, Debug)]
 pub struct PatternDefinition {
     pub name: String,
-    pub(crate) scope: usize,
+    pub scope: usize,
     pub params: Vec<(String, Variable)>,
     // this could just be a usize representing the len
     pub local_vars: Vec<usize>,
@@ -38,62 +33,6 @@ impl PatternDefinition {
             local_vars,
             pattern,
         }
-    }
-
-    pub(crate) fn from_node(
-        node: &Node,
-        context: &CompilationContext,
-        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
-        pattern_definitions: &mut Vec<PatternDefinition>,
-        global_vars: &mut BTreeMap<String, usize>,
-        logs: &mut AnalysisLogs,
-    ) -> Result<()> {
-        // TODO: make sure pattern definitions are only allowed at the top level
-        let name = node
-            .child_by_field_name("name")
-            .ok_or_else(|| anyhow!("missing name of patternDefinition"))?;
-        let name = name.utf8_text(context.src.as_bytes())?;
-        let name = name.trim();
-        let scope_index = vars_array.len();
-        vars_array.push(vec![]);
-        let mut local_vars = BTreeMap::new();
-        // important that this occurs first, as calls assume
-        // that parameters are registered first
-        let params = get_variables(
-            &context
-                .pattern_definition_info
-                .get(name)
-                .ok_or_else(|| anyhow!("cannot get info for pattern {}", name))?
-                .parameters,
-            context.file,
-            vars_array,
-            scope_index,
-            &mut local_vars,
-            global_vars,
-        )?;
-
-        let body = node
-            .child_by_field_name("body")
-            .ok_or_else(|| anyhow!("missing body of patternDefinition"))?;
-        let body = AndCompiler::from_node(
-            &body,
-            context,
-            &mut local_vars,
-            vars_array,
-            scope_index,
-            global_vars,
-            logs,
-        )?;
-        let pattern_def = PatternDefinition::new(
-            name.to_owned(),
-            scope_index,
-            params,
-            local_vars.values().cloned().collect(),
-            body,
-        );
-        // todo check for duplicate names
-        pattern_definitions.push(pattern_def);
-        Ok(())
     }
 
     pub(crate) fn call<'a>(
